+++ CotEditor/Sources/Application/Defaults/DefaultKeys.swift
@@ -87,6 +87,8 @@ extension DefaultKeys {
     static let detectsIndentStyle = DefaultKey<Bool>("detectsIndentStyle")
     static let autoExpandTab = DefaultKey<Bool>("autoExpandTab")
     static let indentWithTabKey = DefaultKey<Bool>("indentWithTabKey")
+    static let insertsCommentDelimitersAfterIndent = DefaultKey<Bool>("insertsCommentDelimitersAfterIndent")
+    static let appendsCommentSpacer = DefaultKey<Bool>("appendsCommentSpacer")
     static let autoLinkDetection = DefaultKey<Bool>("autoLinkDetectionKey")
     static let checkSpellingAsType = DefaultKey<Bool>("checkSpellingAsType")
     static let highlightBraces = DefaultKey<Bool>("highlightBraces")