+++ CotEditor/Sources/Document Window/Content View/EditorTextViewController.swift
@@ -150,6 +150,11 @@ final class EditorTextViewController: NSViewController, NSServicesMenuRequestor,
             defaults.publisher(for: .hangingIndentWidth, initial: true)
                 .assign(to: \.hangingIndentWidth, on: self.textView),
             
+            defaults.publisher(for: .insertsCommentDelimitersAfterIndent, initial: true)
+                .assign(to: \.commentsOutAfterIndent, on: self.textView),
+            defaults.publisher(for: .appendsCommentSpacer, initial: true)
+                .sink { [weak self] in self?.textView.commentSpacer = $0 ? " " : "" },
+            
             defaults.publisher(for: .pageGuideColumn, initial: true)
                 .assign(to: \.pageGuideColumn, on: self.textView),
             defaults.publisher(for: .overscrollRate, initial: true)