+++ CotEditor/Sources/Scripting Support/TextSelection.swift
@@ -239,7 +239,7 @@ private enum OSAUnicodeNormalizationType: String {
     /// Uncomments the selection.
     @objc func handleCommentOut(_ command: NSScriptCommand) {
         
-        self.textView?.commentOut(types: .both, fromLineHead: false)
+        self.textView?.commentOut(types: .both, at: .selection)
     }